moved fixtures from test_masks.py to conftest.py
