+++ tests/test_masks.py
@@ -1,16 +1,7 @@
-import pytest
 from src.masks import get_mask_card_number
 from src.masks import get_mask_account
-
-
-@pytest.fixture
-def correct_card_number():
-    return '4016223056204055'
-
-
-@pytest.fixture
-def correct_account_number():
-    return "2501620256234890001"
+from tests.conftest import correct_account_number
+from tests.conftest import correct_card_number
 
 
 def test_get_mask_card_number(correct_card_number):
